Make SHAKE128 x4 always available and use CPU feature detection
Move contents of sha3x4.c into sha3x4_avx2.c
